Component-ize Footer and add variants
